Refactoring and Corrections
Some refactoring and documentation done in some classes.
Also corrections in the way T1 and T2 values are calculated on the
HashingFunction, to produce even clusters. This is an important
correction because it is a more strict implementation of Distance Based
Hashing, which assures the similarity properties are kept in the
produced buckets (cluster candidates).