Print Original trajectories in Simplified Trajectory Methods
Plot Original Trajectories along with simplified trajectories for
Clusters that need simplified same length (fixed size) vectors to
process such as Kmeans Euclidean and LSH Euclidean.
Presents some bugs (Throws exception) when num of trajectories in
simplified set of cluster does not match the number of trajectories in
the original set of data.