Added support to print Intracluster distance matrix
Added support to print  (and export to a file) the Intracluster distance
matrix so it is easy to figure out the pairwise distance between
elements of a given cluster. This is for comparing items within an
cluster and have a better idea of the approximation results, specially
for approximation generated clusters such as DBH and LSH based clusters.
This will help to figure out the mirror problem observed in the cluster
plots, where some clusters group together trajectories with the same
shape but in another axis in a mirror kind of way.
Resources file updated with the path to export the intracluster distance
matrix.
For now only DTW metric is supported for intracluster distance matrix.
New Class ClusterQualityMeterer will contain many methods to evaluate
cluster quality, in a need to separate concerns from the
testTrajectoryCluster class. Needs more refactoring.