Added Changes to Speed Up DTW with FastDTW
Added some changes to use FastDTW by Stan Salvador, in an attempt to
verify if the speed up provided by any constraint calculation of DTW
give similar clustering results. So far this attemps have shown that
FastDTW is slower than regular DTW, something must be wrong, or maybe
becuase trajectories are not big enough to benefit from the speedup.
Also considering if FastDTW is a good Idea since I only need the DTW
distance and not the Path.
Added Functionality to print Feature Vectors to a File so Zay can read
them better.